+++ src/components/emailSelector/Label.tsx
@@ -33,6 +33,7 @@ interface LabelProps {
   suggestions?: Suggestions;
   renderSuggestion?: React.ComponentType<React.ComponentProps<typeof EmailSuggestion>>;
   menuWidth?: number;
+  disableRemovalOfFirst?: boolean;
 }
 
 const Label = ({
@@ -48,6 +49,7 @@ const Label = ({
   onFinish,
   renderSuggestion,
   menuWidth,
+  disableRemovalOfFirst,
 }: LabelProps) => {
   const [content, setContent] = useState(option.email);
   const [autocompleteOpen, setAutocompleteOpen] = useState(true);
@@ -149,7 +151,7 @@ const Label = ({
           break;
 
         case BACKSPACE:
-          if (!event.repeat && content === '' && onRemove) {
+          if (!event.repeat && content === '' && onRemove && !((index === 0 || index === 1) && disableRemovalOfFirst)) {
             onRemove(index);
           }
           break;
@@ -237,7 +239,7 @@ const Label = ({
       margin={1}
       marginRight={2}
       onFocus={onTagFocus}
-      onRemoveClick={onRemove && onTagRemove}
+      onRemoveClick={!(index === 0 && disableRemovalOfFirst) && onRemove && onTagRemove}
     >
       <Link element="button" className={theme['label-text']} inherit={false} onClick={onTagClick}>
         {option.label || option.email}